Bug fix: was deriving Serialize and Deserialize.
It isn't possible to derive Serialize and Deserialize on any of parking_lot's
types as they are variations on mutexes, which will require us to lock them
before we can do anything with them. Moreover, I had already hand implemented
these; the only reason there wasn't a compiler conflict was because there were
trait bounds that made the implementations appear to be different.  So, all
derivations of Serialize and Deserialize have been removed in favor of the hand
written versions.